Enterprise Support Page updates
Some tweaks to the verbiage around what is provided with Premium Support and GHAE.

Instead of the "Government notification box" indicating the TSAM will inform them of the difference, I propose a secondary article, similar to how we've setup Premium Support, with something like:
For more information, see "GitHub AE Support for Commercial and Government Cloud."
This new page should house all of the roles/responsibilities for each of the two respective services in a table format that's clear and defines the support offering better.

Also, while reviewing, this seems like it should be designated with "Cloud" to denote GHES and GHE**C**:
https://docs.github.com/en/enterprise-server@2.22/admin/enterprise-support/about-github-premium-support-for-github-enterprise
to "About GitHub Premium Support for GitHub Enterprise Cloud"

So we ultimately have:
About GitHub Premium Support for GitHub Enterprise Server
About GitHub Premium Support for GitHub Enterprise Cloud
About GitHub AE Support for Commercial and Government

The navigation would simply show the Commercial/Government Cloud page under "About GitHub Enterprise Support" in the navigation so the customers can see the differences in the two and we can better explain what we provide. Both of the "About Premium Support" pages expand on a lot of good information about Premium Support.