Clarify hotpatch rollback instructions
Replication does not need to be stopped before hotpatching, so most customers faced with a rollback will probably have replication running. Since replication can't be running for a package upgrade (which is how we roll back a hotpatch) we should include a reminder in our docs that replication will need to be stopped prior to rolling back.