Code Scanning: Specify the recommended CodeQL version for each GHES version
Create new variables for the GHES version, and the CodeQL CLI version
recommended along with each GHES version.

Refactor the docs on installing the CodeQL CLI in a third-party CI system,
to use the CodeQL and GHES version from the variables.

These variables will need to be kept up to date with future GHES+CodeQL versions.