Improve "Configuring a subdomain" section
Current version has two different `www` subdomain examples, but I think the intention was for second subdomain example to be a non-`www` case. Looks like it was supposed to be `another` used as the subdomain in the second example, but instead the domain name was changed instead (`www.example.com` becoming `www.anotherexample.com`).