Fix a typo (I think) in git clone progress text
I compared the git clone progress text with the one in my terminal when I ran git clone on my fork of Spoon-Knife, in the way the instructions indicate. It appears that the "> remote ... remote ... remove: Total n (delta m), reused a (delta b)" contains a typo, with 'remove' instead of 'remote'. I fixed this in case it is a typo.