Update "Telling Git about your X.509 key" to not assume that GPG and X.509 are the only signing options
GitHub has recently introduced a third way to sign your commits
alongside GPG and X.509: using an SSH key.

At the moment, our copy about X.509 mentions using it "instead of
GPG", which made sense when the only options were GPG or it, but
now it's a bit confusing. This removes that copy.