Change the ssh key default scope to github.com
With the current host wildcard default configuration, the new key applies to all ssh connection attempts regardless of host. This can become an issue on older systems that leverage existing keys in the ssh config and don't have knowledge of the "newer" recommended ed25519 algorithm. While this key can be reused for multiple purposes outside of github, I still propose to scope the default host to github.com to avoid a default that can pollute the global key namespace in the ssh config.